requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
kafka-python==2.0.2
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
